move fixRelativePaths to wshcmd-file-util
